Transit Update
useSigner hook is not being called at the right time. The useSigner hook is a reactive hook, which means that it will only update the value of signer when the component re-renders. This means that if the component is not re-rendering, the value of signer will not be updated.

i've tried adding a state variable id and using the useEffect hook to update its value when the query parameter changes